+++ src/main/java/com/grupo5e/morapack/service/PedidoService.java
@@ -7,13 +7,13 @@ import java.util.List;
 
 public interface PedidoService {
     List<Pedido> listar();
-    Long insertar(Pedido pedido);
-    Pedido actualizar(Long id, Pedido pedido);
-    Pedido buscarPorId(Long id);
+    Integer insertar(Pedido pedido);
+    Pedido actualizar(Integer id, Pedido pedido);
+    Pedido buscarPorId(Integer id);
     List<Pedido> buscarPorCliente(Long clienteId);
     List<Pedido> buscarPorEstado(EstadoPedido estado);
-    Pedido actualizarEstado(Long id, EstadoPedido nuevoEstado);
-    void eliminar(Long id);
-    boolean existePorId(Long id);
+    Pedido actualizarEstado(Integer id, EstadoPedido nuevoEstado);
+    void eliminar(Integer id);
+    boolean existePorId(Integer id);
     List<Pedido> insertarBulk(List<Pedido> pedidos);
 }
