+++ src/main/java/com/grupo5e/morapack/service/impl/AeropuertoServiceImpl.java
@@ -33,13 +33,13 @@ public class AeropuertoServiceImpl implements AeropuertoService {
 
     @Override
     @Transactional
-    public Long insertar(Aeropuerto aeropuerto) {
+    public Integer insertar(Aeropuerto aeropuerto) {
         return aeropuertoRepository.save(aeropuerto).getId();
     }
 
     @Override
     @Transactional
-    public Aeropuerto actualizar(Long id, Aeropuerto aeropuerto) {
+    public Aeropuerto actualizar(Integer id, Aeropuerto aeropuerto) {
         Aeropuerto existente = buscarPorId(id);
         if (existente == null) {
             throw new ResourceNotFoundException("Aeropuerto", "id", id);
@@ -50,7 +50,7 @@ public class AeropuertoServiceImpl implements AeropuertoService {
 
     @Override
     @Transactional
-    public Aeropuerto toggleEstado(Long id) {
+    public Aeropuerto toggleEstado(Integer id) {
         Aeropuerto aeropuerto = buscarPorId(id);
         if (aeropuerto == null) {
             throw new ResourceNotFoundException("Aeropuerto", "id", id);
@@ -67,7 +67,7 @@ public class AeropuertoServiceImpl implements AeropuertoService {
     }
 
     @Override
-    public Aeropuerto buscarPorId(Long id) {
+    public Aeropuerto buscarPorId(Integer id) {
         return aeropuertoRepository.findById(id).orElse(null);
     }
 
@@ -78,7 +78,7 @@ public class AeropuertoServiceImpl implements AeropuertoService {
 
     @Override
     @Transactional
-    public void eliminar(Long id) {
+    public void eliminar(Integer id) {
         if (!existePorId(id)) {
             throw new ResourceNotFoundException("Aeropuerto", "id", id);
         }
@@ -86,7 +86,7 @@ public class AeropuertoServiceImpl implements AeropuertoService {
     }
 
     @Override
-    public boolean existePorId(Long id) {
+    public boolean existePorId(Integer id) {
         return aeropuertoRepository.existsById(id);
     }
